feat!: major refactor — module split, security, logging, performance
BREAKING CHANGE: proxy.py (4297 lines) split into 3 modules.
The monolithic proxy.py has been decomposed into:
  - proxy.py    (~2100 lines) — HTTP server, routing, auth, SSE write
  - transformer.py (~2680 lines) — protocol conversion, content, tools, SSE parsing
  - logger.py   (~189 lines) — orjson wrappers, log level control, file rotation

Any code that 'from proxy import <transformer_function>' will break.
All transformer functions must now be imported from the transformer module.

---

## Security (NEW)

- auth_key: API key authentication via Authorization: Bearer or x-api-key header
  (config.json auth_key field, empty = no auth)
- IP blacklist (denied_ips): block specific IPs/CIDR, highest priority
- IP whitelist (allowed_ips): allow only specified IPs/CIDR
- Optional HTTPS: ssl_cert + ssl_key auto-enables TLS (ssl.SSLContext)

## Logging (REDESIGNED)

- New log level system: -1=disabled / 0=silent / 1=error / 2=info(default) / 3=verbose
- New functions: log_error(), log_info(), log_debug() (replaces bare log())
- Priority: config.json log_level > SHTU_LOG_LEVEL env > default 2
- Log file rotation: auto-rename to .log.1 when exceeding 5MB
- register_active_config() callback to avoid logger<->proxy circular import

## Performance

- orjson acceleration (2-10x faster JSON, fallback to stdlib)
- SSE batch write: 6-10 events merged into 1 write+flush
- Shallow copy replaces deepcopy
- str.join replaces string concatenation (+=)
- Lazy visible_text generation
- CJK character token estimation
- OrderedDict preserves insertion order

## Bug Fixes

- stop_reason_from_done: finish_reason 'length' now maps to 'end_turn'
  (was 'max_tokens' which caused Claude Code to misinterpret as truncation)
- Responses API status 'incomplete' also maps to 'end_turn'
- max_tokens auto-clamped when input+output exceeds max_context_tokens
  (fixes deepseek-pro 128k overflow: 96001+32000 > 128000)
- Passthrough mode: non-genaiapi URLs forward SSE directly, no conversion
- Client disconnect: ConnectionResetError/BrokenPipeError/ConnectionAbortedError
  silently suppressed in write_sse/write_sse_batch/write_data_sse/send_json/handle()
- ACTIVE_CONFIG module-level initialization (was missing, caused NameError)
- emit_redacted_thinking_sse moved from transformer to proxy (was calling
  write_sse across module boundary, caused NameError)
- json_dumps_compact duplicate removed from transformer (was referencing
  _orjson raw module, caused NameError)
- _orjson.JSONDecodeError replaced with ValueError (was referencing
  _orjson raw module across boundary)
- register_active_config() call added in main() (was missing, caused
  log_level=2 to produce no output)

## config_store.py

- New AppConfig fields: auth_key, log_level, ssl_cert, ssl_key,
  allowed_ips (List[str]), denied_ips (List[str])

## Files Changed

- src/proxy.py       : MODIFIED (complete rewrite — HTTP layer only)
- src/transformer.py  : NEW (all protocol conversion extracted)
- src/logger.py       : NEW (orjson wrappers + log system)
- src/config_store.py : MODIFIED (+6 fields in AppConfig)
- src/config.json     : NOT tracked (.gitignore)
- README.md           : NOT modified (per user request)
